Move source process cases to Group A
